Struct skala_engine_numerics::vector::Vec3

pub struct Vec3 {
    pub x: SignedFractional,
    pub y: SignedFractional,
    pub z: SignedFractional,
}

A 3d vector.

Fields
x: SignedFractionaly: SignedFractionalz: SignedFractional
Implementations

impl Vec3

pub const ZERO: Self = Self { x: SignedFractional::ZERO, y: SignedFractional::ZERO, z: SignedFractional::ZERO, }

A vector of length zero

pub fn new<A, B, C>(x: A, y: B, z: C) -> Self where
A: Into<SignedFractional>,
B: Into<SignedFractional>,
C: Into<SignedFractional>,

Creates a new Vec3 from coordinates

Examples

let pos = Vec3::new(1, 1i32 , 5u8);

pub fn magnitude_pow2(&self) -> SignedFractional

Returns the magnitude of this Vec3 raised to the power of two.

Examples

let x = Vec3::new(1, 0, 0);

// Proving we're working with a unit vector
assert_eq!(x.magnitude_pow2(), 1);

pub fn magnitude(&self) -> SignedFractional

Returns the magnitude of this Vec3

Examples

let x = Vec3::new(2, 4, 4);

assert_eq!(x.magnitude(), 6);

pub fn normalize(&mut self)

Sets the magnitude of this Vec3 to one

Panics

If vector is magnitude is zero

Examples

let mut x = Vec3::new(20, 0, 0);

// Before normalization
assert_eq!(x.magnitude(), 20);

x.normalize();
// After normalization
assert_eq!(x.magnitude(), 1);

pub fn get_normalized(&self) -> Self

Creates a Vec3 with magnitude equal to one and rotation equal to this Vec3

Panics

If vector magnitude is 0

Examples

let x = Vec3::new(10, 0, 0);

assert_eq!(x.get_normalized(), Vec3::new(1, 0, 0));

pub fn try_get_normalized(&self) -> Option<Self>

Creates a Vec3 with magnitude equal to one and rotation equal to this Vec3

Examples

let x = Vec3::new(10, 0, 0);
let zero = Vec3::new(0, 0, 0);

assert_eq!(x.try_get_normalized(), Some(Vec3::new(1, 0, 0)));
assert_eq!(zero.try_get_normalized(), None);